Extract plugin-api-worldstate module
Move the seven world state observation contracts (WorldStateService,
TrieLogService and the trielogs package) into a new
ethereum/core/plugin-api-worldstate module published as
besu-plugin-api-worldstate, re-exported by besu-plugin-api so the
umbrella surface is unchanged. Relocate WorldStateServiceImpl from
app/services to the ethereum.worldstate.pluginadapter package in
ethereum/core, following the owning-subsystem convention. The four
backend contracts stay in the umbrella per the split agreed on #11085.

Issue #11085
