+++ ethereum/core/plugin-api-worldstate/build.gradle
@@ -1,0 +1,60 @@
+/*
+ * Copyright contributors to Besu.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with
+ * the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on
+ * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations under the License.
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+apply plugin: 'java-library'
+
+// override the path-derived name (besu-ethereum-core-plugin-api-worldstate) with the plugin API family naming
+def worldstateApiArtifactId = 'besu-plugin-api-worldstate'
+
+jar {
+  archiveBaseName = worldstateApiArtifactId
+  manifest {
+    attributes(
+      'Specification-Title': archiveBaseName,
+      'Specification-Version': project.version,
+      'Implementation-Title': archiveBaseName,
+      'Implementation-Version': calculateVersion(),
+      'Commit-Hash': getGitCommitDetails(40).hash,
+      'Automatic-Module-Name': 'org.hyperledger.besu.plugin.api.worldstate'
+      )
+  }
+}
+
+dependencies {
+  api project(':plugin-api:core')
+  // AccountValue, Address, Hash and StorageSlotKey appear in this module's own
+  // signatures, so the dependency is declared here even though plugin-api:core
+  // also re-exports it
+  api project(':datatypes')
+  // WorldStateService returns evm.worldstate.WorldView from both of its methods;
+  // this is the module's only evm edge
+  api project(':evm')
+  // Bytes and UInt256 appear in the trie-log contract signatures
+  api 'io.consensys.tuweni:tuweni-bytes'
+  api 'io.consensys.tuweni:tuweni-units'
+}
+
+publishing {
+  publications {
+    mavenJava(MavenPublication) {
+      groupId = 'org.hyperledger.besu'
+      artifactId = worldstateApiArtifactId
+      pom {
+        name = 'Besu Plugin API Worldstate'
+        description = 'World state observation contracts of the Besu plugin API, letting plugins read world state and observe trie log changes'
+      }
+    }
+  }
+}
