+++ platform/build.gradle
@@ -60,6 +60,7 @@ dependencies {
     api project(':ethereum:blockcreation')
     api project(':ethereum:core')
     api project(':ethereum:core:plugin-api-chain')
+    api project(':ethereum:core:plugin-api-worldstate')
     api project(':ethereum:eth')
     api project(':ethereum:eth:plugin-api-sync')
     api project(':ethereum:eth:plugin-api-txpool')
