+++ plugin-api/build.gradle
@@ -33,6 +33,7 @@ dependencies {
   api project(':crypto:plugin-api')
   api project(':ethereum:permissioning:plugin-api')
   api project(':services:kvstore:plugin-api')
+  api project(':ethereum:core:plugin-api-worldstate')
   api project(':ethereum:p2p:plugin-api')
   api project(':ethereum:eth:plugin-api-txpool')
   api project(':ethereum:eth:plugin-api-sync')
@@ -70,6 +71,7 @@ tasks.register('checkAPICompatibility', me.champeau.gradle.japicmp.JapicmpTask) 
     project(':ethereum:permissioning:plugin-api').tasks.named('jar'),
     project(':crypto:plugin-api').tasks.named('jar'),
     project(':services:kvstore:plugin-api').tasks.named('jar'),
+    project(':ethereum:core:plugin-api-worldstate').tasks.named('jar'),
     project(':ethereum:p2p:plugin-api').tasks.named('jar'),
     project(':ethereum:eth:plugin-api-txpool').tasks.named('jar'),
     project(':ethereum:eth:plugin-api-sync').tasks.named('jar'),
