+++ settings.gradle
@@ -44,6 +44,7 @@ include 'ethereum:api:plugin-api'
 include 'ethereum:blockcreation'
 include 'ethereum:core'
 include 'ethereum:core:plugin-api-chain'
+include 'ethereum:core:plugin-api-worldstate'
 include 'ethereum:eth'
 include 'ethereum:eth:plugin-api-sync'
 include 'ethereum:eth:plugin-api-txpool'
